requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
greenlet==3.2.3
idna==3.10
psycopg2-binary==2.9.10
pyTelegramBotAPI==4.27.0
python-dotenv==1.1.0
requests==2.32.3
SQLAlchemy==2.0.41
typing_extensions==4.14.0
urllib3==2.4.0
